+++ src/main/scala/org/podval/tools/publish/markup/EntityName.scala
@@ -4,7 +4,7 @@ import org.podval.xml.XmlCodec
 import zio.blocks.schema.{Modifier, Schema}
 
 final case class EntityName(
-//  kind: EntityKind,
+  kind: EntityKind,
   @Modifier.config(XmlCodec.Attribute, "") id: Option[String] = None,
   @Modifier.config(XmlCodec.Attribute, "") ref: Option[String] = None,
   @Modifier.config(XmlCodec.Text, "") name: String
@@ -12,4 +12,4 @@ final case class EntityName(
 
 object EntityName:
   given schema: Schema[EntityName] = Schema.derived
-  val codec: XmlCodec[EntityName] = XmlCodec.derived
+  val codec: XmlCodec[EntityName] = XmlCodec.derived[EntityName, EntityKind]("kind", EntityKind.asName)